Move settings to dedicated page and drop possession view
