+++ src/components/Dashboard.tsx
@@ -2,7 +2,6 @@ import React, { useState, useRef, useEffect } from 'react';
 import { GameState } from '../types';
 import { GamePresetSelector } from './GamePresetSelector';
 import { getHalfName } from '../utils/gamePresets';
-import { ConfirmModal } from './ConfirmModal';
 import { MatchSummary } from './MatchSummary';
 import {
   Play,
@@ -13,7 +12,6 @@ import {
   Upload,
   Monitor,
   BarChart3,
-  Timer,
   Undo2,
   Redo2,
   Settings,
@@ -23,55 +21,45 @@ import { useSettings } from '../hooks/useSettings';
 interface DashboardProps {
   gameState: GameState;
   updateTeam: (team: 'home' | 'away', field: 'name' | 'score' | 'fouls' | 'logo', value: string | number) => void;
-  updateTournamentLogo: (logo: string) => void;
-  updateTournamentName: (name: string) => void;
   updateTime: (minutes: number, seconds: number) => void;
   toggleTimer: () => void;
   resetTimer: () => void;
   updatePeriod: (period: number) => void;
   changeGamePreset: (presetIndex: number) => void;
-  resetGame: (options?: { force?: boolean }) => void;
   undo: () => void;
   redo: () => void;
   addPlayer: (team: 'home' | 'away', name: string) => void;
   removePlayer: (team: 'home' | 'away', playerId: string) => void;
   onViewChange: (
-    view: 'scoreboard' | 'dashboard' | 'overlay' | 'stats' | 'possession' | 'settings'
+    view: 'scoreboard' | 'dashboard' | 'overlay' | 'stats' | 'settings'
   ) => void;
 }
 
 export const Dashboard: React.FC<DashboardProps> = ({
   gameState,
   updateTeam,
-  updateTournamentLogo,
-  updateTournamentName,
   updateTime,
   toggleTimer,
   resetTimer,
   updatePeriod,
   changeGamePreset,
-  resetGame,
   undo,
   redo,
   addPlayer,
   removePlayer,
   onViewChange,
 }) => {
-  const [activeTab, setActiveTab] = useState<'teams' | 'timer' | 'format' | 'settings'>('teams');
-  const tabs = ['teams', 'timer', 'format', 'settings'] as const;
+  const [activeTab, setActiveTab] = useState<'teams' | 'timer' | 'format'>('teams');
+  const tabs = ['teams', 'timer', 'format'] as const;
   const { settings } = useSettings();
 
   const [homeLogoError, setHomeLogoError] = useState('');
   const [awayLogoError, setAwayLogoError] = useState('');
-  const [tournamentLogoError, setTournamentLogoError] = useState('');
   const [homePlayerName, setHomePlayerName] = useState('');
   const [awayPlayerName, setAwayPlayerName] = useState('');
 
-  const [showResetConfirm, setShowResetConfirm] = useState(false);
-
   const homeLogoUrlRef = useRef<string | null>(null);
   const awayLogoUrlRef = useRef<string | null>(null);
-  const tournamentLogoUrlRef = useRef<string | null>(null);
 
   const homePlayers = gameState.homeTeam.players ?? [];
   const awayPlayers = gameState.awayTeam.players ?? [];
@@ -105,37 +93,15 @@ export const Dashboard: React.FC<DashboardProps> = ({
     }
   };
 
-  const handleTournamentLogoUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (!file) return;
-
-    if (!file.type.startsWith('image/') || file.type === 'image/svg+xml') {
-      setTournamentLogoError('Please upload a valid image (SVG not allowed).');
-      return;
-    }
-
-    const url = URL.createObjectURL(file);
-    if (tournamentLogoUrlRef.current) {
-      URL.revokeObjectURL(tournamentLogoUrlRef.current);
-    }
-    tournamentLogoUrlRef.current = url;
-    updateTournamentLogo(url);
-    setTournamentLogoError('');
-  };
-
   useEffect(() => {
     const homeUrl = homeLogoUrlRef.current;
     const awayUrl = awayLogoUrlRef.current;
-    const tournamentUrl = tournamentLogoUrlRef.current;
     return () => {
       if (homeUrl) {
         URL.revokeObjectURL(homeUrl);
       }
       if (awayUrl) {
         URL.revokeObjectURL(awayUrl);
-      }
-      if (tournamentUrl) {
-        URL.revokeObjectURL(tournamentUrl);
       }
     };
   }, []);
@@ -180,33 +146,6 @@ export const Dashboard: React.FC<DashboardProps> = ({
     }
   };
 
-  const handleResetGame = () => {
-    const hasStarted =
-      gameState.isRunning ||
-      gameState.time.minutes !== gameState.gamePreset.halfDuration ||
-      gameState.time.seconds !== 0 ||
-      gameState.homeTeam.score !== 0 ||
-      gameState.awayTeam.score !== 0 ||
-      gameState.homeTeam.fouls !== 0 ||
-      gameState.awayTeam.fouls !== 0 ||
-      Object.values(gameState.homeTeam.stats).some(v => v !== 0) ||
-      Object.values(gameState.awayTeam.stats).some(v => v !== 0);
-
-    if (!hasStarted) {
-      resetGame({ force: true });
-    } else {
-      setShowResetConfirm(true);
-    }
-  };
-
-  const confirmResetGame = () => {
-    resetGame({ force: true });
-    setShowResetConfirm(false);
-  };
-
-  const cancelResetGame = () => {
-    setShowResetConfirm(false);
-  };
 
   return (
     <div className="min-h-screen bg-gray-50 text-gray-900 dark:bg-gray-900 dark:text-gray-100">
@@ -236,13 +175,6 @@ export const Dashboard: React.FC<DashboardProps> = ({
               >
                 <BarChart3 className="w-4 h-4" />
                 Stats Tracker
-              </button>
-              <button
-                onClick={() => onViewChange('possession')}
-                className="inline-flex items-center gap-2 px-4 py-2 bg-orange-600 text-white rounded-lg hover:bg-orange-700 transition-colors"
-              >
-                <Timer className="w-4 h-4" />
-                Possession Control
               </button>
               <button
                 onClick={() => onViewChange('settings')}
@@ -679,104 +611,7 @@ export const Dashboard: React.FC<DashboardProps> = ({
           />
         )}
 
-        {/* Settings Tab */}
-        {activeTab === 'settings' && (
-          <div className="bg-white dark:bg-gray-800 rounded-xl shadow-sm border border-gray-200 dark:border-gray-700 p-8 max-w-2xl mx-auto">
-            <h3 className="text-lg font-semibold text-gray-900 dark:text-gray-100 mb-8 text-center">Game Settings</h3>
-            
-            <div className="space-y-8">
-              {/* Tournament Settings */}
-              <div className="space-y-6">
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">Tournament Name</label>
-                  <input
-                    type="text"
-                    value={gameState.tournamentName}
-                    onChange={e => updateTournamentName(e.target.value)}
-                    className="w-full px-3 py-2 border border-gray-300 dark:border-gray-600 rounded-lg focus:ring-2 focus:ring-purple-500 focus:border-purple-500 dark:bg-gray-700 dark:text-gray-100"
-                  />
-                </div>
-                <div>
-                  <label className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-2">Tournament Logo</label>
-                  <div className="flex items-center gap-6">
-                    <div className="w-20 h-20 bg-gray-100 dark:bg-gray-700 rounded-lg overflow-hidden flex items-center justify-center border-2 border-gray-300 dark:border-gray-600">
-                      {gameState.tournamentLogo ? (
-                        <img src={gameState.tournamentLogo} alt="Tournament" className="w-full h-full object-contain" />
-                      ) : (
-                        <span className="text-gray-400 text-xs text-center">No Logo</span>
-                      )}
-                    </div>
-                    <div className="flex gap-3">
-                      <label className="cursor-pointer inline-flex items-center gap-2 px-4 py-2 bg-gray-100 text-gray-700 rounded-lg hover:bg-gray-200 transition-colors dark:bg-gray-700 dark:text-gray-200 dark:hover:bg-gray-600">
-                        <Upload className="w-4 h-4" />
-                        Upload Logo
-                        <input type="file" accept="image/*" onChange={handleTournamentLogoUpload} className="hidden" />
-                      </label>
-                      {gameState.tournamentLogo && (
-                        <button
-                          onClick={() => updateTournamentLogo('')}
-                          className="px-4 py-2 bg-red-100 text-red-700 rounded-lg hover:bg-red-200 transition-colors dark:bg-red-900 dark:text-red-200 dark:hover:bg-red-800"
-                        >
-                          Remove Logo
-                        </button>
-                      )}
-                    </div>
-                  </div>
-                  {tournamentLogoError && (
-                    <p className="text-sm text-red-600 dark:text-red-400 mt-2">{tournamentLogoError}</p>
-                  )}
-                </div>
-              </div>
-
-                <div className="text-center">
-                <button
-                  onClick={handleResetGame}
-                  className="inline-flex items-center gap-2 px-6 py-3 bg-red-600 text-white rounded-lg hover:bg-red-700 transition-colors font-semibold"
-                >
-                  <RotateCcw className="w-5 h-5" />
-                  Reset Entire Game
-                </button>
-                <p className="text-sm text-gray-500 dark:text-gray-400 mt-2">
-                  This will reset scores, fouls, timer, and all settings to default values.
-                </p>
-              </div>
-
-              <div className="border-t pt-6">
-                <h4 className="font-semibold text-gray-900 dark:text-gray-100 mb-4">Quick Actions</h4>
-                <div className="grid grid-cols-2 gap-4">
-                  <button
-                    onClick={() => {
-                      updateTeam('home', 'score', 0);
-                      updateTeam('away', 'score', 0);
-                    }}
-                    className="px-4 py-2 bg-gray-100 text-gray-700 rounded-lg hover:bg-gray-200 transition-colors dark:bg-gray-700 dark:text-gray-200 dark:hover:bg-gray-600"
-                  >
-                    Reset Scores
-                  </button>
-                  <button
-                    onClick={() => {
-                      updateTeam('home', 'fouls', 0);
-                      updateTeam('away', 'fouls', 0);
-                    }}
-                    className="px-4 py-2 bg-gray-100 text-gray-700 rounded-lg hover:bg-gray-200 transition-colors dark:bg-gray-700 dark:text-gray-200 dark:hover:bg-gray-600"
-                  >
-                    Reset Fouls
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
       </div>
-      <ConfirmModal
-        isOpen={showResetConfirm}
-        title="Reset Game"
-        message="Are you sure you want to reset the entire game?"
-        confirmText="Reset"
-        cancelText="Cancel"
-        onConfirm={confirmResetGame}
-        onCancel={cancelResetGame}
-      />
     </div>
   );
 };
