+++ src/hooks/SettingsProvider.tsx
@@ -1,7 +1,12 @@
 import { useState, useEffect, ReactNode } from 'react';
 import { SettingsContext, Settings } from './SettingsContext';
 
-const defaultSettings: Settings = { showUndo: true, showRedo: true };
+const defaultSettings: Settings = {
+  showUndo: true,
+  showRedo: true,
+  homeShortcut: 'KeyA',
+  awayShortcut: 'KeyD',
+};
 
 interface SettingsProviderProps {
   children: ReactNode;
@@ -29,9 +34,13 @@ export const SettingsProvider = ({ children }: SettingsProviderProps) => {
 
   const toggleUndo = () => setSettings(s => ({ ...s, showUndo: !s.showUndo }));
   const toggleRedo = () => setSettings(s => ({ ...s, showRedo: !s.showRedo }));
+  const setHomeShortcut = (code: string) => setSettings(s => ({ ...s, homeShortcut: code }));
+  const setAwayShortcut = (code: string) => setSettings(s => ({ ...s, awayShortcut: code }));
 
   return (
-    <SettingsContext.Provider value={{ settings, toggleUndo, toggleRedo }}>
+    <SettingsContext.Provider
+      value={{ settings, toggleUndo, toggleRedo, setHomeShortcut, setAwayShortcut }}
+    >
       {children}
     </SettingsContext.Provider>
   );
